requirements.txt
PyYAML>=6.0
requests>=2.31.0
psutil>=5.9.0
pytz>=2023.3